move pow folding tests to test_schedule [pr]
not really belongs to test_const_folding
